Make Query methods publicly available
Make the Query methods public to allow for them to be called outside of the base class.
Also add a QueryPassword method.
